--- react/Greeting.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-type Props = {
-  name: string
-}
-
-function Greeting({ name }: Props) {
-  return <div>Heyssss, {name}</div>
-}
-
-export default Greeting
