--- react/Teste.tsx
@@ -1,7 +1,0 @@
-import React, { FC } from 'react'
-
-const Teste: FC = () => {
-  return <div>Teste</div>
-}
-
-export default Teste
