+++ react/components/PromotionsShelf/index.tsx
@@ -32,6 +32,9 @@ const PromotionsShelf: StorefrontFunctionComponent<PromotionsShelfProps> = ({
     error: promotionsError,
   } = usePromotions(promotionId)
 
+  // eslint-disable-next-line no-console
+  console.log('🚀 ~ PromotionsShelf ~ promotions:', promotions)
+
   const allPromotions = promotionParser(promotions)
 
   if (promotionsError || promotionsLoading || !allPromotions.length) {
