+++ react/utils/getFacetFromPromotions.ts
