+++ react/utils/promotionsParser.ts
